Move component TelephoneAdvisoryReportsNumberCard from 'components/cards' to 'components/index/CardsReference'
